Fixed a bunch of things related to ps vita

+++ modules/d_std/source/core/stdc/stddef.d
@@ -1,4 +1,4 @@
 module core.stdc.stddef;
 
-alias c_long = int;
+alias c_long = ptrdiff_t;
 alias wchar_t = wchar;
